move determineNextVersion to core
